docs: better english
"Whether or not" is incorrect, even though it is understandable, the language is incorrect and can just be whether, as whether infers there is a negative and a positive, https://english.stackexchange.com/questions/3382/whether-or-not-vs-whether. In the case of this, "or not" in "Whether or not" is not needed, as it is always undetermined the response in the documentation, the documentation won't know.